Register document/image/embed rich text handlers via hooks
