+++ wagtail/wagtailcore/rich_text.py
@@ -4,15 +4,6 @@ from django.utils.html import escape
 
 from wagtail.wagtailcore.whitelist import Whitelister
 from wagtail.wagtailcore.models import Page
-
-from wagtail.wagtaildocs.models import Document
-
-# FIXME: we don't really want to import wagtailimages within core.
-# For that matter, we probably don't want core to be concerned about translating
-# HTML for the benefit of the hallo.js editor...
-from wagtail.wagtailimages.models import get_image_model
-from wagtail.wagtailimages.formats import get_image_format
-
 from wagtail.wagtailcore import hooks
 
 
@@ -21,80 +12,6 @@ from wagtail.wagtailcore import hooks
 # verbatim in the DB representation because they embed information which is stored
 # elsewhere in the database and is liable to change - from real HTML representation
 # to DB representation and back again.
-
-class ImageEmbedHandler(object):
-    """
-    ImageEmbedHandler will be invoked whenever we encounter an element in HTML content
-    with an attribute of data-embedtype="image". The resulting element in the database
-    representation will be:
-    <embed embedtype="image" id="42" format="thumb" alt="some custom alt text">
-    """
-    @staticmethod
-    def get_db_attributes(tag):
-        """
-        Given a tag that we've identified as an image embed (because it has a
-        data-embedtype="image" attribute), return a dict of the attributes we should
-        have on the resulting <embed> element.
-        """
-        return {
-            'id': tag['data-id'],
-            'format': tag['data-format'],
-            'alt': tag['data-alt'],
-        }
-
-    @staticmethod
-    def expand_db_attributes(attrs, for_editor):
-        """
-        Given a dict of attributes from the <embed> tag, return the real HTML
-        representation.
-        """
-        Image = get_image_model()
-        try:
-            image = Image.objects.get(id=attrs['id'])
-            format = get_image_format(attrs['format'])
-
-            if for_editor:
-                try:
-                    return format.image_to_editor_html(image, attrs['alt'])
-                except:
-                    return ''
-            else:
-                return format.image_to_html(image, attrs['alt'])
-
-        except Image.DoesNotExist:
-            return "<img>"
-
-
-class MediaEmbedHandler(object):
-    """
-    MediaEmbedHandler will be invoked whenever we encounter an element in HTML content
-    with an attribute of data-embedtype="media". The resulting element in the database
-    representation will be:
-    <embed embedtype="media" url="http://vimeo.com/XXXXX">
-    """
-    @staticmethod
-    def get_db_attributes(tag):
-        """
-        Given a tag that we've identified as a media embed (because it has a
-        data-embedtype="media" attribute), return a dict of the attributes we should
-        have on the resulting <embed> element.
-        """
-        return {
-            'url': tag['data-url'],
-        }
-
-    @staticmethod
-    def expand_db_attributes(attrs, for_editor):
-        """
-        Given a dict of attributes from the <embed> tag, return the real HTML
-        representation.
-        """
-        from wagtail.wagtailembeds import format
-        if for_editor:
-            return format.embed_to_editor_html(attrs['url'])
-        else:
-            return format.embed_to_frontend_html(attrs['url'])
-
 
 class PageLinkHandler(object):
     """
@@ -127,33 +44,9 @@ class PageLinkHandler(object):
             return "<a>"
 
 
-class DocumentLinkHandler(object):
-    @staticmethod
-    def get_db_attributes(tag):
-        return {'id': tag['data-id']}
-
-    @staticmethod
-    def expand_db_attributes(attrs, for_editor):
-        try:
-            doc = Document.objects.get(id=attrs['id'])
-
-            if for_editor:
-                editor_attrs = 'data-linktype="document" data-id="%d" ' % doc.id
-            else:
-                editor_attrs = ''
-
-            return '<a %shref="%s">' % (editor_attrs, escape(doc.url))
-        except Document.DoesNotExist:
-            return "<a>"
-
-
-EMBED_HANDLERS = {
-    'image': ImageEmbedHandler,
-    'media': MediaEmbedHandler,
-}
+EMBED_HANDLERS = {}
 LINK_HANDLERS = {
     'page': PageLinkHandler,
-    'document': DocumentLinkHandler,
 }
 
 has_loaded_embed_handlers = False
